+++ source/include/utils.h
@@ -126,7 +126,7 @@ namespace utils {
         return "";
     }
 
-    std::string toBinary(int n)
+    std::string toBinary(long n)
     {
         std::string r;
         while(n!=0) {r=(n%2==0 ?"0":"1")+r; n/=2;}
